fix(clauderon): change UploadResponse size to u32 for TypeShare compatibility
TypeShare only supports u32 for unsigned integers, not u64 or usize.
Since max file size is 10MB (well under u32's 4GB limit), u32 is sufficient.

+++ packages/clauderon/src/api/http_server.rs
@@ -526,7 +526,7 @@ async fn upload_file(
 
     Ok(Json(crate::api::protocol::UploadResponse {
         path: file_path.to_string_lossy().to_string(),
-        size: file_data.len() as u64,
+        size: file_data.len() as u32,
     }))
 }
